Exercise 2 almost done
