+++ burger_shop/HealthBurger.java
@@ -1,63 +1,40 @@
 package burger_shop;
 
 public class HealthBurger extends Burger{
-	private double healthPrice = 7.99;
-	public HealthBurger(String name, String breadType, String meat) {
-		super("Healthy", breadType, meat);
-		
-	}
-	
-	
-	@Override
-	public double hamburguerPrice(double One) {
-		
-		
-		this.healthPrice = this.healthPrice+One;
-		System.out.println("Hamburguer price: $" + this.healthPrice);
-		return this.healthPrice;
-		
-	}
-	@Override
-	public double hamburguerPrice(double One,double Two) {
-		this.healthPrice = this.healthPrice+One+Two;
-		System.out.println("Hamburguer price: $" + this.healthPrice);
-		return this.healthPrice;
-		
-	}
-	@Override
-	public double hamburguerPrice(double One,double Two,double Three) {
-		this.healthPrice = this.healthPrice+One+Two+Three;
-		System.out.println("Hamburguer price: $" + this.healthPrice);
-		return this.healthPrice;
-		
-	}
-	@Override
-	public double hamburguerPrice(double One,double Two,double Three,double Four) {
-		
-		this.healthPrice = this.healthPrice+One+Two+Three+Four;
-		System.out.println("Hamburguer price: $" + this.healthPrice);
-		return this.healthPrice;
-		
-	}
-	@Override
-	public double hamburguerPrice(double One,double Two, double Three, double Four,double Five) {
-		if(getName().contains("Health")) {
-			System.out.println("You can only have 4 toppings for a plain burguer");
-		}
-		
-		return this.healthPrice ;
-		
-	}
-	@Override
-	public double hamburguerPrice(double One,double Two, double Three, double Four,double Five,double Six) {
-		if(getName().contains("Health")) {
-			System.out.println("You can only have 4 toppings for a plain burguer");
-		}
-		
-		return this.healthPrice ;
-		
-		
-	}
-	
-	
+
+    private String healthyExtra1Name;
+    private double healthyExtra1Price;
+
+    private String healthyExtra2Name;
+    private double healthyExtra2Price;
+
+    public HealthBurger(String meat, double price) {
+        super("Healthy", meat, price, "Brown rye");
+    }
+
+    public void addHealthAddition1(String name, double price) {
+        this.healthyExtra1Name = name;
+        this.healthyExtra1Price = price;
+    }
+
+    public void addHealthAddition2(String name, double price) {
+        this.healthyExtra2Name = name;
+        this.healthyExtra2Price = price;
+    }
+
+    @Override
+    public double itemizeHamburger() {
+        double hamburgerPrice = super.itemizeHamburger();
+        if(this.healthyExtra1Name != null) {
+            hamburgerPrice += this.healthyExtra1Price;
+            System.out.println("Added " + this.healthyExtra1Name + " for an extra " + this.healthyExtra1Price);
+        }
+        if(this.healthyExtra2Name != null) {
+            hamburgerPrice += this.healthyExtra2Price;
+            System.out.println("Added " + this.healthyExtra2Name + " for an extra " + this.healthyExtra2Price);
+        }
+
+        return hamburgerPrice;
+    }
+
 }
