+++ burger_shop/Order.java
@@ -1,4 +1,5 @@
 package burger_shop;
 
 public class Order {
+
 }
